+++ libft/ft_strncmp.c
@@ -6,11 +6,13 @@
 /*   By: phan <phan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 14:09:59 by phan              #+#    #+#             */
-/*   Updated: 2023/03/18 14:57:03 by phan             ###   ########.fr       */
+/*   Updated: 2023/03/18 20:17:01 by phan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned int	i;
