+++ libft/ft_strnstr.c
@@ -6,11 +6,13 @@
 /*   By: phan <phan@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 16:49:41 by phan              #+#    #+#             */
-/*   Updated: 2023/03/18 17:51:13 by phan             ###   ########.fr       */
+/*   Updated: 2023/03/18 20:17:23 by phan             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_strnstr(const char *haystack, const char *needle, unsigned int len)
+#include "libft.h"
+
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
 	unsigned int	h_idx;
 	unsigned int	n_idx;
